requirements.txt
accelerate>=0.21.0
av>=11.0.0
clip @ https://github.com/openai/CLIP/archive/d50d76daa670286dd6cacf3bcd80b5e4823fc8e1.zip#sha256=b5842c25da441d6c581b53a5c60e0c2127ebafe0f746f8e15561a006c6c3be6a
decord>=0.6.0
diffusers>=0.24.0
einops>=0.4.1
imageio>=2.33.0
imageio-ffmpeg>=0.4.9
numpy>=1.23.5
omegaconf>=2.2.3
onnxruntime>=1.16.3
onnxruntime-gpu>=1.16.3
open-clip-torch>=2.20.0
opencv-contrib-python>=4.8.1.78
opencv-python>=4.8.1.78
Pillow>=9.5.0
scikit-image>=0.21.0
scikit-learn>=1.3.2
scipy>=1.11.4
torchdiffeq>=0.2.3
torchmetrics>=1.2.1
torchsde>=0.2.5
tqdm>=4.66.1
transformers>=4.30.2